requirements.txt
pyside6
ffmpeg-python
